make class to file
