+++ Assets/Scripts/Enemy/Enemy.cs
@@ -140,96 +140,6 @@ namespace Enemy
         }
     }
 
-    public class AwayState : IEnemyState
-    {
-        public void OnEnter(Enemy enemy)
-        {
-            enemy.navMeshAgentWrapper.SetSpeed(NavMeshAgentWrapper.MoveSpeed.Run);
-            var leaveRange = 10f;
-            var hatePosition = enemy.targetEnemy.transform.position;
-            var transform = enemy.transform;
-            var position = transform.position;
-            enemy.navMeshAgentWrapper.SetDestination(position +
-                                                     (position - hatePosition).normalized * leaveRange);
-        }
 
-        public void OnExit(Enemy enemy)
-        {
-            enemy.navMeshAgentWrapper.SetSpeed(NavMeshAgentWrapper.MoveSpeed.Walk);
-        }
 
-        public void OnUpdate(Enemy enemy)
-        {
-            var enemyState = Enemy.EnemyState.Away;
-            if (enemy.navMeshAgentWrapper.IsArrived())
-            {
-                enemyState = Enemy.EnemyState.Idle;
-            }
-
-            enemy.SetState(enemyState);
-        }
-    }
-
-    public class AttackState : IEnemyState
-    {
-        public void OnEnter(Enemy enemy)
-        {
-            enemy.navMeshAgentWrapper.SetDestination(enemy.transform.position);
-            enemy.weaponCollider.enabled = true;
-            enemy.weaponCollider.isTrigger = true;
-            enemy.weapon.Reset();
-        }
-
-        public void OnExit(Enemy enemy)
-        {
-            enemy.weaponCollider.enabled = false;
-            enemy.navMeshAgentWrapper.SetSpeed(NavMeshAgentWrapper.MoveSpeed.Run);
-        }
-
-        public void OnUpdate(Enemy enemy)
-        {
-            var enemyState = Enemy.EnemyState.Attack;
-            if (enemy.animator.GetCurrentAnimatorStateInfo(0).IsName("Attack") &&
-                enemy.animator.GetCurrentAnimatorStateInfo(0).normalizedTime > 1)
-            {
-                enemyState = Enemy.EnemyState.Away;
-            }
-
-            enemy.SetState(enemyState);
-        }
-    }
-
-    public class ChaseState : IEnemyState
-    {
-        public void OnEnter(Enemy enemy)
-        {
-            // animator.Play("Chase");
-            enemy.navMeshAgentWrapper.SetSpeed(NavMeshAgentWrapper.MoveSpeed.Sprint);
-        }
-
-        public void OnExit(Enemy enemy)
-        {
-            enemy.weaponCollider.enabled = false;
-            enemy.navMeshAgentWrapper.SetSpeed(NavMeshAgentWrapper.MoveSpeed.Run);
-        }
-
-        public void OnUpdate(Enemy enemy)
-        {
-            enemy.SetDestination(enemy.targetEnemy.transform.position);
-            var enemyState = Enemy.EnemyState.Chase;
-            if (Vector3.Distance(enemy.transform.position, enemy.targetEnemy.transform.position) <
-                enemy.GetAttackRange())
-            {
-                enemyState = Enemy.EnemyState.Attack;
-            }
-
-            if (Vector3.Distance(enemy.transform.position, enemy.targetEnemy.transform.position) > 4f &&
-                !enemy.navMeshAgentWrapper.Visible(enemy.targetEnemy.gameObject))
-            {
-                enemyState = Enemy.EnemyState.Idle;
-            }
-
-            enemy.SetState(enemyState);
-        }
-    }
 }
