+++ Assets/Scripts/Enemy/enemyState/AwayState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 0003773158654ba2a5e7cb321febf3f3
+timeCreated: 1735476317
